Add form

+++ admin/models/marca.php
@@ -3,9 +3,10 @@ require_once __DIR__ . '/../model.php';
 
 class Marca extends Model
 {
-    function create()
+    function create($data)
     {
-
+        print_r($data);
+        return $data;
     }
 
     function update()
@@ -19,6 +20,18 @@ class Marca extends Model
         $this->conn->beginTransaction();
         $count = 0;
         try {
+            $sql = "SELECT COUNT(*) as count FROM producto WHERE id_marca = :id_marca";
+            $data = $this->conn->prepare($sql);
+            $data->bindParam(':id_marca', $id, PDO::PARAM_INT);
+            $data->execute();
+            $count = $data->fetch(PDO::FETCH_ASSOC);
+            $count = $count['count'];
+
+            if ($count > 0) {
+                $this->conn->rollBack();
+                return false;
+            }
+
             $sql = "DELETE FROM marca WHERE id_marca = :id_marca";
             $data = $this->conn->prepare($sql);
             $data->bindParam(':id_marca', $id, PDO::PARAM_INT);
